chore: Move transpilation to generators
